requirements.txt
BeautifulSoup==3.2.1
Django==1.6.5
Fabric==1.9.0
GitPython==0.3.2.RC1
M2Crypto==0.22.3
MySQL-python==1.2.5
Pillow==2.5.3
PyPDF2==1.23
South==1.0
Unipath==1.0
Werkzeug==0.9.6
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
async==0.6.1
billiard==3.3.0.18
celery==3.1.16
django-dajax==0.9.2
django-dajaxice==0.6
django-databrowse==1.3
django-extensions==1.3.8
django-model-report==0.2.1
django-storages==1.1.8
django-tastypie==0.11.1
ecdsa==0.11
factory-boy==2.4.1
future==0.15.1
geo-utils==0.1
geopy==1.0.1
gitdb==0.5.4
gitflow==0.5.1
html5lib==0.999
ipython==2.2.0
kombu==3.0.24
logger==1.4
nose==1.3.3
numpy==1.8.1
oauthlib==1.0.3
paramiko==1.14.0
pycrypto==2.6.1
# pycups==1.9.67
# pyodbc==3.0.7
python-dateutil==2.2
python-memcached==1.53
python-mimeparse==0.1.4
pytz==2014.9
reportlab==3.1.8
requests==2.7.0
requests-oauthlib==0.5.0
six==1.7.3
smmap==0.8.2
wsgiref==0.1.2
xhtml2pdf==0.0.6
xlwt==0.7.4
git+https://github.com/erikvw/django-crypto-fields@master#egg=django_crypto_fields
git+https://github.com/erikvw/django-revision@0.1.8#egg=django_revision
git+https://github.com/botswana-harvard/edc.git@1.13.21#egg=edc
git+https://github.com/botswana-harvard/edc-audit@0.1.0#egg=edc_audit
git+https://github.com/botswana-harvard/edc-admin-exclude.git@0.1.3#egg=edc_admin_exclude
git+https://github.com/botswana-harvard/edc-base@0.1.8#egg=edc_base
git+https://github.com/botswana-harvard/edc-consent@0.1.13#egg=edc_consent
git+https://github.com/botswana-harvard/edc-constants@0.1.6#egg=edc_constants
git+https://github.com/botswana-harvard/edc-device.git@0.1.3#egg=edc_device
git+https://github.com/botswana-harvard/edc-quota@1.0.11#egg=edc_quota
git+https://github.com/botswana-harvard/getresults-identifier@0.1.0#egg=getresults_identifier
git+https://github.com/botswana-harvard/lis@1.7.3#egg=lis
